added sidebar toggle button
-different components will be rendered based on what button is clicked
